data-warehouse-runner/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Naked==0.1.32
odps==3.5.1
pycryptodome==3.18.0
PyMySQL==1.1.0
pyodps==0.11.4.1
PyYAML==6.0
requests==2.31.0
shellescape==3.8.1
SQLAlchemy==2.0.18
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.6
